+++ packages/ra-core/src/controller/field/index.ts
@@ -1,34 +1,3 @@
-import ReferenceArrayFieldController, {
-    ReferenceArrayFieldControllerProps,
-} from './ReferenceArrayFieldController';
-import ReferenceFieldController, {
-    ReferenceFieldControllerProps,
-} from './ReferenceFieldController';
-import ReferenceManyFieldController, {
-    ReferenceManyFieldControllerProps,
-} from './ReferenceManyFieldController';
-import getResourceLinkPath, { LinkToType } from './getResourceLinkPath';
-import useReferenceArrayFieldController from './useReferenceArrayFieldController';
-import useReferenceManyFieldController from './useReferenceManyFieldController';
-import { ListControllerProps } from '../list';
-
-// kept for backwards compatibility
-// deprecated to be removed in 4.0
-export type ReferenceArrayProps = ListControllerProps;
-export type ReferenceManyProps = ListControllerProps;
-
-export type {
-    LinkToType,
-    ReferenceArrayFieldControllerProps,
-    ReferenceFieldControllerProps,
-    ReferenceManyFieldControllerProps,
-};
-
-export {
-    useReferenceArrayFieldController,
-    ReferenceArrayFieldController,
-    ReferenceFieldController,
-    getResourceLinkPath,
-    useReferenceManyFieldController,
-    ReferenceManyFieldController,
-};
+export * from './getResourceLinkPath';
+export * from './useReferenceArrayFieldController';
+export * from './useReferenceManyFieldController';
